requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
pytesseract>=0.3.10
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pandas>=2.0.0
